requirements.txt
Django==2.2
django-crispy-forms==1.9.1
djangorestframework==3.11.0
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
